requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
# Add packages specifically for Vercel deployment
requests==2.31.0
urllib3==2.0.7